Add owner seed and fix owner migrate Fix getAllOwners method in ownerDAO Related to #6 Signed-off-by: João Paulo Ferreira <iamferreirajp@gmail.com> Signed-off-by: Adailson Santos <adailson2@gmail.com>

+++ src/main/resources/db/migration/V1.1.1__Create_owner_table.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS `owner`;
 CREATE TABLE `owner`(
-	`cpf` int(10) NOT NULL,
+	`cpf` bigint(11) NOT NULL,
 	`ownerName` varchar(50) NOT NULL,
 	`ownerLastName` varchar(50) NOT NULL,
-	`phoneNumber` int(20) NOT NULL,
+	`phoneNumber` bigint(20) NOT NULL,
  	`address` varchar(200) NOT NULL,
  	PRIMARY KEY(`cpf`)
 );
